student_dubt_project/requirements.txt
fastapi
pydantic
langchain_google_genai
langchain
langchain_core
uvicorn
wikipedia
google-generativeai
